Created first release version of my Todolist project v.1

+++ TodoList/RESTService/Scripts/script.js
@@ -29,21 +29,50 @@
 
         $scope.submitNewTask = function () {
 
-            $scope.isVisiblePopup = false;
-            var data = { "title": $scope.title, "description": $scope.description };
+            if ($scope.id) {
 
-            $http.post(
-                '/api/task',
-                JSON.stringify(data),
-                {
-                    headers: { 'Content-Type': 'application/json' }
-                }
-            ).then(function (response) {
-                $scope.tasks.push(response.data);
-            },
-                function () {
-                    console.log("Something went wrong. New task wasn't added.");
-                });
+                // Update
+
+                $scope.tasks[$scope.id].Title = $scope.title;
+                $scope.tasks[$scope.id].Description = $scope.description;
+
+                var data = { "id": $scope.id, "title": $scope.title, "description": $scope.description, "publishDate": $scope.tasks[$scope.id].PublishDate };
+
+                $http.put(
+                    '/api/task',
+                    JSON.stringify(data),
+                    {
+                        headers: { 'Content-Type': 'application/json' }
+                    }
+                ).then(function (response) {
+                    $scope.tasks.push(response.data);
+                },
+                    function () {
+                        console.log("Something went wrong. The new task with id: " + $scope.id + " wasn't added.");
+                    });
+            }
+            else {
+
+                // Create
+
+                var data = { "title": $scope.title, "description": $scope.description };
+
+                $http.post(
+                    '/api/task',
+                    JSON.stringify(data),
+                    {
+                        headers: { 'Content-Type': 'application/json' }
+                    }
+                ).then(function (response) {
+                    $scope.tasks.push(response.data);
+                },
+                    function () {
+                        console.log("Something went wrong. New task wasn't added.");
+                    });
+            }
+
+            $scope.isVisiblePopup = false;
+
         };
 
         $scope.getQueryResults = function () {
@@ -61,23 +90,38 @@
                 });
         };
 
-        $scope.deleteTask = function(index) {
+        $scope.deleteTask = function (index) {
 
-            console.log("Into delete task function.");
-            console.log("Index = " + index);
-            console.log($scope.tasks[index]);
-
-            $http.delete(
-                '/api/task',
-                {
-                    params: { id: $scope.tasks[index].Id }
-                }
-            ).then(function (response) {
-                $scope.tasks.splice(index, 1);
-            },
+            if (confirm("Do you really want to delete this task?")) {
+                $http.delete(
+                    '/api/task',
+                    {
+                        params: { id: $scope.tasks[index].Id }
+                    }
+                ).then(function (response) {
+                    $scope.tasks.splice(index, 1);
+                },
                 function () {
                     console.log("Something went wrong. The was't deleted.");
                 });
+            }
+
+        };
+
+        $scope.editTask = function (index) {
+            $scope.title = $scope.tasks[index].Title;
+            $scope.description = $scope.tasks[index].Description;
+            $scope.id = $scope.tasks[index].Id;
+
+            $scope.isVisiblePopup = true;
+        };
+
+        $scope.createTask = function () {
+            $scope.title = '';
+            $scope.description = '';
+            $scope.id = 0;
+
+            $scope.isVisiblePopup = true;
         };
 
     }]);
